Add files via upload
Added ancillary files that are useful to the project, but not actually a part of making WV models.
Added "sic" to the possible tags for models. This is supposed to have the model get its vocabulary from descriptions of SIC groups, but more generally allows the model to use an arbitrary vocabulary not based on the corpus.
Changed sims.py to allow default values for the diagonal.
Minor variable changes to Autotest.py and plots.py from use.
Added a first draft readme to make the project easier to use.